backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=21.0
psycopg[binary]>=3.1
dj-database-url>=2.1
whitenoise>=6.6
python-dotenv